ksqldb - fix protobuf examples
