+++ ksqldb/protobuf-without-schema-registry/proto-python-client/requirements.txt
@@ -1,3 +1,3 @@
 confluent-kafka>=1.4.2
-protobuf==3.11.2
+protobuf==4.24.4
 six==1.14.0
